requirements.txt
langchain==0.2.6
langchain-community==0.2.6
streamlit==1.36.0
python-docx==1.1.2
pypandoc==1.13
odfpy==1.4.1
PyPDF2==3.0.1
faiss-cpu==1.8.0.post1